requirements.txt
flask
python-telegram-bot
openai
requests
